refactor(Filter): move the value-control dispatch into its own module
`valueControl` and `placeholder` lived in Filter.vue's <script setup>, which
made them unreachable to any other component and untestable: the dispatch
imports `.vue` value inputs and `frappe-ui`, and frappe-ui is a peer dependency
with no copy installed in this package, so a module importing it cannot be
loaded by the test runner at all.

Split so the rules are pure and the components are a lookup:

- valueControl.ts returns a ValueControlId plus props — no component imports,
  so it loads under vitest
- valueControlComponents.ts maps each id to its component
- Filter.vue renders VALUE_CONTROLS[valueControl(f).control]

No behaviour change: the dispatch order, the props and the placeholder copy are
carried over verbatim. The nested ConditionBuilder will render the same value
inputs from this module rather than growing a second copy of the rules.

Adds 24 tests over the dispatch, which had none — covering the precedence that
matters (operator-driven swaps beat fieldtype, so a control is never handed a
value it cannot represent) and the in/not-in option-field split.
