+++ ui/src/components/Filter/Filter.vue
@@ -9,14 +9,11 @@
   `crm.api.doc.get_filterable_fields` + `getOperators` (ADR-0001, ADR-0003).
 
   The value inputs are the shared `Fields` module's components (ADR-0004), not CRM's
-  bespoke Link/Duration/Rating trio. The gaps the shared inputs don't cover are
-  operator-driven swaps handled here, in the `.vue`: `is`/`is not` → a Set/Not-Set
-  select, `timespan` → a preset select, `like`/`not like` → a plain text box, and
-  Date `between` → a range picker. `in`/`not in` over an option field (Select /
-  Autocomplete / Link) is a MultiSelect of the field's values (`MultiSelectInput` /
-  `MultiLinkInput`) so values are picked, not typed as an error-prone comma string;
-  over a free-text field it stays a comma text box. Field pickers use frappe-ui's
-  `Combobox`; icons are lucide names.
+  bespoke Link/Duration/Rating trio. Which control a condition gets — the
+  operator-driven swaps as well as the fieldtype ones — is decided by
+  `valueControl.ts`, a sibling module rather than this file, so the nested
+  ConditionBuilder renders the same inputs from the same rules. Field pickers use
+  frappe-ui's `Combobox`; icons are lucide names.
 -->
 <template>
 	<!-- Empty: a plain "Filter" button that opens the field picker (the first
@@ -100,7 +97,7 @@
 							</div>
 							<div class="w-[180px]">
 								<component
-									:is="valueControl(f).is"
+									:is="VALUE_CONTROLS[valueControl(f).control]"
 									v-bind="valueControl(f).props"
 									class="w-full"
 									:modelValue="f.value"
@@ -153,33 +150,17 @@
 
 <script setup lang="ts">
 import { computed, nextTick, ref } from "vue";
-import { Button, Combobox, Popover, Select, TextInput, DateRangePicker } from "frappe-ui";
+import { Button, Combobox, Popover, Select } from "frappe-ui";
 import { useDoctypeMeta } from "../../composables/useDoctypeMeta";
 import { getFilterableFields } from "./getFilterableFields";
-import {
-	getOperators,
-	conditionFor,
-	carryOver,
-	defaultValueFor,
-	isOptionField,
-} from "./operators";
+import { getOperators, conditionFor, carryOver, defaultValueFor } from "./operators";
 import type { Filter, FilterField, FilterOperator, FilterProps, FilterValue } from "./types";
-// `in` / `not in` over an option field picks values from a MultiSelect rather than
-// typing a comma string: a static list for Select/Autocomplete, a live link search
-// for Link.
-import MultiSelectInput from "./MultiSelectInput.vue";
-import MultiLinkInput from "./MultiLinkInput.vue";
-// The shared, fieldtype-aware value inputs (ADR-0004). Filter mounts only the
-// zero-coupling subset; it never provides the form-context injections, so the
-// deep-injection inputs (Number resolving currency) fall back to site defaults.
-import SelectField from "../Fields/SelectField.vue";
-import LinkField from "../Fields/LinkField.vue";
-import NumberField from "../Fields/NumberField.vue";
-import DateField from "../Fields/DateField.vue";
-import DatetimeField from "../Fields/DatetimeField.vue";
-import DurationField from "../Fields/DurationField.vue";
-import RatingField from "../Fields/RatingField.vue";
-import type { FieldMeta } from "../Fields/types";
+// The operator/fieldtype → value-input dispatch. Lives in its own module so the
+// nested ConditionBuilder can render the same value inputs from the same rules
+// without duplicating them; it returns an id, and VALUE_CONTROLS maps that id to
+// the component (see valueControl.ts for why the two are separate files).
+import { valueControl } from "./valueControl";
+import { VALUE_CONTROLS } from "./valueControlComponents";
 
 const props = defineProps<FilterProps>();
 
@@ -248,145 +229,5 @@ function removeFilter(index: number) {
 function clearAll(close: () => void) {
 	model.value = [];
 	close();
-}
-
-// --- Value-control dispatch -------------------------------------------------
-// Mirrors CRM's `getValueControl`: operator-driven swaps first, then fieldtype.
-// Fieldtype controls are the shared `Fields` components; the swaps reuse frappe-ui
-// primitives. (Type groups are re-declared here, as CRM does, so `operators.ts`
-// stays a pure helper.)
-const CHECK_TYPES = ["Check"];
-const LINK_TYPES = ["Link", "Dynamic Link"];
-const NUMBER_TYPES = ["Float", "Int", "Currency", "Percent"];
-const SELECT_TYPES = ["Select", "Autocomplete"];
-const DATE_TYPES = ["Date", "Datetime"];
-const DURATION_TYPES = ["Duration"];
-const RATING_TYPES = ["Rating"];
-const TEXT_OPERATORS = ["like", "not like", "in", "not in"];
-
-const SET_OPTIONS = [
-	{ label: "Set", value: "set" },
-	{ label: "Not Set", value: "not set" },
-];
-
-const TIMESPAN_OPTIONS = [
-	"last week",
-	"last month",
-	"last quarter",
-	"last 6 months",
-	"last year",
-	"yesterday",
-	"today",
-	"tomorrow",
-	"this week",
-	"this month",
-	"this quarter",
-	"this year",
-	"next week",
-	"next month",
-	"next quarter",
-	"next 6 months",
-	"next year",
-].map((v) => ({ label: titleCase(v), value: v }));
-
-function titleCase(s: string): string {
-	return s.replace(/\b\w/g, (c) => c.toUpperCase());
-}
-
-interface ValueControl {
-	is: unknown;
-	props: Record<string, unknown>;
-}
-
-/** Field meta passed to a `Fields` value input — no label/description so the
- *  control renders bare inside the compact filter row. */
-function bareField(field: FilterField, overrides: Partial<FieldMeta> = {}): FieldMeta {
-	return {
-		fieldname: field.fieldname,
-		fieldtype: field.fieldtype,
-		options: field.options,
-		...overrides,
-	};
-}
-
-function valueControl(f: Filter): ValueControl {
-	const operator = f.operator;
-	const field = f.field;
-	const fieldtype = field?.fieldtype ?? "Data";
-	const ph = placeholder(f);
-
-	// Operator-driven swaps (handled here, not in the field components).
-	if (operator === "is" || operator === "is not") {
-		return { is: Select, props: { options: SET_OPTIONS, placeholder: ph } };
-	}
-	if (operator === "timespan") {
-		return { is: Select, props: { options: TIMESPAN_OPTIONS, placeholder: ph } };
-	}
-	// `in` / `not in` over an option field → a MultiSelect of the field's values
-	// (picked, not typed comma-separated). Link searches its target doctype; Select
-	// and Autocomplete read their options from meta. Dynamic Link and free-text
-	// fields fall through to the comma TextInput below.
-	if ((operator === "in" || operator === "not in") && isOptionField(fieldtype)) {
-		if (LINK_TYPES.includes(fieldtype)) {
-			return { is: MultiLinkInput, props: { field: field! } };
-		}
-		return { is: MultiSelectInput, props: { field: field! } };
-	}
-	if (TEXT_OPERATORS.includes(operator)) {
-		return { is: TextInput, props: { type: "text", placeholder: ph } };
-	}
-
-	// Fieldtype-driven value inputs from the shared Fields module.
-	if (SELECT_TYPES.includes(fieldtype) || CHECK_TYPES.includes(fieldtype)) {
-		// Check filters on Yes/No; Select on its own meta options.
-		const options = CHECK_TYPES.includes(fieldtype) ? "Yes\nNo" : field?.options;
-		return {
-			is: SelectField,
-			props: { field: bareField(field!, { options, placeholder: ph }) },
-		};
-	}
-	if (LINK_TYPES.includes(fieldtype)) {
-		// Dynamic Link has no fixed target doctype to pick against — plain text.
-		if (fieldtype === "Dynamic Link")
-			return { is: TextInput, props: { type: "text", placeholder: ph } };
-		return { is: LinkField, props: { field: bareField(field!, { placeholder: ph }) } };
-	}
-	if (NUMBER_TYPES.includes(fieldtype)) {
-		return { is: NumberField, props: { field: bareField(field!, { placeholder: ph }) } };
-	}
-	if (DATE_TYPES.includes(fieldtype) && operator === "between") {
-		return { is: DateRangePicker, props: { iconLeft: "" } };
-	}
-	if (DURATION_TYPES.includes(fieldtype)) {
-		return { is: DurationField, props: { field: bareField(field!, { placeholder: ph }) } };
-	}
-	if (RATING_TYPES.includes(fieldtype)) {
-		return { is: RatingField, props: { field: bareField(field!) } };
-	}
-	if (DATE_TYPES.includes(fieldtype)) {
-		const is = fieldtype === "Date" ? DateField : DatetimeField;
-		return { is, props: { field: bareField(field!, { placeholder: ph }) } };
-	}
-	return { is: TextInput, props: { type: "text", placeholder: ph } };
-}
-
-/** Per-operator / per-fieldtype placeholder copy. A port of CRM's `placeholder`,
- *  except `like`/`not like` show a bare term (not `%John%`): `serializeFilters`
- *  wraps the value in `%` itself, so prompting for the wildcards would mislead. */
-function placeholder(f: Filter): string {
-	const fieldtype = f.field?.fieldtype ?? "Data";
-	if (f.operator === "between") return "01/01/2022 to 01/31/2022";
-	if (f.operator === "in" || f.operator === "not in")
-		return NUMBER_TYPES.includes(fieldtype) ? "100, 200, 300" : "John, Jane, Doe";
-	if (f.operator === "like" || f.operator === "not like")
-		return NUMBER_TYPES.includes(fieldtype) ? "100" : "John";
-	if (f.operator === "is" || f.operator === "is not") return "Set";
-	if (f.operator === "timespan") return "Last Week";
-	if (NUMBER_TYPES.includes(fieldtype)) return "1000";
-	if (DATE_TYPES.includes(fieldtype)) return "01/01/2022";
-	if (CHECK_TYPES.includes(fieldtype)) return "Yes";
-	if (LINK_TYPES.includes(fieldtype)) return "Select a Value";
-	if (SELECT_TYPES.includes(fieldtype)) return "Select an Option";
-	return "John Doe";
 }
 </script>
